add/remove time fragments
a range [start, end] can be added or excluded from the tracked time fragments of a task,
essentially to join fragments or to ”cut a hole" in the time covered by a particular task.
This can be used to add forgotten time or to remove wrongly associated time or
to "move" tracked time from one task to another.