UI for adding/removing time fragments
context menu entries to start/stop tracking, add or remove complete fragments.
popup dialog to enter start and end of a time fragment to be added/removed.

loading / quitting prompts a dialog that asks if the currently modified document should be saved.
the dialog has a cancel option that prevents the application from quitting.